requirements.txt
alembic==1.10.2
asgiref==3.5.2
asyncpg==0.24.0
click==8.1.3
fastapi==0.88.0
greenlet==1.1.2
h11==0.13.0
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.1
pydantic==1.10.7
PyGreSQL==5.2.4
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
typer==0.4.1
typing_extensions==4.5.0
uvicorn==0.21.1
zipp==3.8.0
cachetools==5.3.0
Authlib==1.2.0
itsdangerous==2.1.2
httpx==0.23.3
pyjwt==2.7.0
requests==2.30.0
geoip2